backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
mozilla-django-oidc==4.0.1
python-jose==3.3.0
requests==2.31.0
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0
